Moved asset class to proper file
